Change ClientManager so that clients stay in the database even after they log out.
THere's also a updateClientToDatabase method, which should be called internally to the ClientManager to update clients when their state is modified (such as a location update).

NodeManager is getting ready to be integrated with the database as well.

The Node object no longer stores an id, nodes that have ids will be a DatabaseNode instead.